Add files via upload
Code for setting the initial conditions so that shallow_water_simulation.py can calculate the shape of the surface of the water over the duration of the simulation. It also creates and shows the resulting animation of the water and whether or not the simulation is stable for the given time step dt.