Added animations
Added the animations. Included the original animation, as well as a ver2 (version 2) which uses a custom version of the 'Blues' cmap from the matplotlib package to colour the surface of the water. This new version has better shading than the original version so the waves are clearer. Doesn't look very 3D in the new one though so need to work on that.